Barrelized components
Updated all components, and pages using those components, to barrelize
them and to use more appropriate names (capitalized).
