+++ src/components/Alert/Alert.scss
