+++ src/components/Alert/index.js
@@ -1,0 +1,1 @@
+export { Alert as default } from './Alert';
