+++ src/components/Code/index.js
@@ -1,0 +1,1 @@
+export { Code as default } from './Code';
