+++ src/components/ExternalLink/ExternalLink.js
@@ -22,5 +22,3 @@ ExternalLink.propTypes = {
     href: PropTypes.string.isRequired,
     children: PropTypes.node.isRequired
 };
-
-export default ExternalLink;
