+++ src/components/ExternalLink/index.js
@@ -1,0 +1,1 @@
+export { ExternalLink as default } from './ExternalLink';
