+++ src/components/Icon/index.js
@@ -1,0 +1,1 @@
+export { Icon as default } from './Icon';
