+++ src/components/SEO/SEO.js
@@ -98,5 +98,3 @@ SEO.defaultProps = {
     meta: [],
     description: ''
 };
-
-export default SEO;
