+++ src/components/SEO/index.js
@@ -1,0 +1,1 @@
+export { SEO as default } from './SEO';
