+++ src/components/ScrollToTop/index.js
@@ -1,0 +1,1 @@
+export { ScrollToTop as default } from './ScrollToTop';
