+++ src/pages/utilities.js
@@ -1,9 +1,8 @@
 import React from 'react';
-
-import Alert from '../components/alert';
-import Code from '../components/code';
-import Layout from '../components/layout';
-import SEO from '../components/seo';
+import Alert from '../components/Alert';
+import Code from '../components/Code';
+import Layout from '../components/Layout';
+import SEO from '../components/SEO';
 
 const title = 'Utility classes',
     description =
@@ -16,7 +15,7 @@ const Summary = () => (
     </>
 );
 
-const IndexPage = () => (
+const UtilitiesPage = () => (
     <Layout summary={Summary}>
         <SEO title={title} description={description} />
         <h2>Text alignment</h2>
@@ -306,4 +305,4 @@ const IndexPage = () => (
     </Layout>
 );
 
-export default IndexPage;
+export default UtilitiesPage;
